Rename logging exporter to debug exporter
