+++ examples/local/otel-config.yaml
@@ -32,7 +32,7 @@ processors:
   batch:
 
 exporters:
-  logging:
+  debug:
     logLevel: debug
 
 service:
@@ -40,10 +40,10 @@ service:
     traces:
       receivers: [otlp, opencensus, jaeger, zipkin]
       processors: [batch]
-      exporters: [logging]
+      exporters: [debug]
     metrics:
       receivers: [otlp, opencensus, prometheus]
       processors: [batch]
-      exporters: [logging]
+      exporters: [debug]
 
   extensions: [health_check, pprof, zpages]
