+++ exporter/README.md
@@ -29,7 +29,7 @@ Available log exporters (sorted alphabetically):
 Available local exporters (sorted alphabetically):
 
 - [File](fileexporter/README.md)
-- [Logging](loggingexporter/README.md)
+- [Debug](debugexporter/README.md)
 
 The [contrib
 repository](https://github.com/open-telemetry/opentelemetry-collector-contrib)
