+++ exporter/debugexporter/debug_exporter.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package loggingexporter
+package debugexporter
 
 import (
 	"context"
@@ -29,7 +29,7 @@ import (
 	"go.opentelemetry.io/collector/model/pdata"
 )
 
-type loggingExporter struct {
+type debugExporter struct {
 	logger           *zap.Logger
 	debug            bool
 	logsMarshaler    pdata.LogsMarshaler
@@ -37,7 +37,7 @@ type loggingExporter struct {
 	tracesMarshaler  pdata.TracesMarshaler
 }
 
-func (s *loggingExporter) pushTraces(_ context.Context, td pdata.Traces) error {
+func (s *debugExporter) pushTraces(_ context.Context, td pdata.Traces) error {
 	s.logger.Info("TracesExporter", zap.Int("#spans", td.SpanCount()))
 
 	if !s.debug {
@@ -52,7 +52,7 @@ func (s *loggingExporter) pushTraces(_ context.Context, td pdata.Traces) error {
 	return nil
 }
 
-func (s *loggingExporter) pushMetrics(_ context.Context, md pdata.Metrics) error {
+func (s *debugExporter) pushMetrics(_ context.Context, md pdata.Metrics) error {
 	s.logger.Info("MetricsExporter", zap.Int("#metrics", md.MetricCount()))
 
 	if !s.debug {
@@ -67,7 +67,7 @@ func (s *loggingExporter) pushMetrics(_ context.Context, md pdata.Metrics) error
 	return nil
 }
 
-func (s *loggingExporter) pushLogs(_ context.Context, ld pdata.Logs) error {
+func (s *debugExporter) pushLogs(_ context.Context, ld pdata.Logs) error {
 	s.logger.Info("LogsExporter", zap.Int("#logs", ld.LogRecordCount()))
 
 	if !s.debug {
@@ -82,8 +82,8 @@ func (s *loggingExporter) pushLogs(_ context.Context, ld pdata.Logs) error {
 	return nil
 }
 
-func newLoggingExporter(level string, logger *zap.Logger) *loggingExporter {
-	return &loggingExporter{
+func newDebugExporter(level string, logger *zap.Logger) *debugExporter {
+	return &debugExporter{
 		debug:            strings.ToLower(level) == "debug",
 		logger:           logger,
 		logsMarshaler:    otlptext.NewTextLogsMarshaler(),
@@ -95,7 +95,7 @@ func newLoggingExporter(level string, logger *zap.Logger) *loggingExporter {
 // newTracesExporter creates an exporter.TracesExporter that just drops the
 // received data and logs debugging messages.
 func newTracesExporter(config config.Exporter, level string, logger *zap.Logger, set component.ExporterCreateSettings) (component.TracesExporter, error) {
-	s := newLoggingExporter(level, logger)
+	s := newDebugExporter(level, logger)
 	return exporterhelper.NewTracesExporter(
 		config,
 		set,
@@ -112,7 +112,7 @@ func newTracesExporter(config config.Exporter, level string, logger *zap.Logger,
 // newMetricsExporter creates an exporter.MetricsExporter that just drops the
 // received data and logs debugging messages.
 func newMetricsExporter(config config.Exporter, level string, logger *zap.Logger, set component.ExporterCreateSettings) (component.MetricsExporter, error) {
-	s := newLoggingExporter(level, logger)
+	s := newDebugExporter(level, logger)
 	return exporterhelper.NewMetricsExporter(
 		config,
 		set,
@@ -129,7 +129,7 @@ func newMetricsExporter(config config.Exporter, level string, logger *zap.Logger
 // newLogsExporter creates an exporter.LogsExporter that just drops the
 // received data and logs debugging messages.
 func newLogsExporter(config config.Exporter, level string, logger *zap.Logger, set component.ExporterCreateSettings) (component.LogsExporter, error) {
-	s := newLoggingExporter(level, logger)
+	s := newDebugExporter(level, logger)
 	return exporterhelper.NewLogsExporter(
 		config,
 		set,
